create quiz command now updates the database and creates a text channel under the quizzes category. Also educator can assign sections to the quiz.
