+++ Python/82bd821df430a70f075996882d7ba59118acf34cc07b584bb37092c67f0959bb.json
@@ -1,0 +1,1 @@
+[{"question": "2+2", "answer": "4", "wrong": ["5", " 6", " 7"], "points": 10.0}, {"question": "2+2", "answer": "4", "wrong": ["5", " 6", " 7"], "points": 10.0}]
