+++ Python/af1d6edc4740497335f46663f20236b4d8069a2f4e9b49709498b88d1ca3c1ab.json
@@ -1,0 +1,1 @@
+[{"question": "2+2", "answer": "4", "wrong": ["5", " 6", " 7"], "points": 20.0}]
