+++ Python/temp.json
@@ -1,0 +1,1 @@
+[{"question": "2+2", "answer": "4", "wrong": ["5", " 6", " 7"], "points": 20.0}]
